requirements.txt
Django==1.8.3
django-bootstrap3==6.1.0
django-debug-toolbar==1.3.2
django-registration-redux==1.1
django-social-auth==0.7.28
django-tagging==0.4
httplib2==0.9.1
oauth2==1.5.211
oauthlib==1.0.1
Pillow==2.9.0
PyJWT==1.4.0
python-openid==2.2.5
python-social-auth==0.2.12
redis==2.10.3
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
sorl-thumbnail==12.3
sqlparse==0.1.16
Wand==0.4.1
Unidecode==0.4.18
django-friendship==1.3.0
